refactor(UserDao): AddStatement 클래스를 add()에 로컬 클래스로 이동
